+++ examples/simpletrigger/simpletrigger/processor.py
@@ -10,7 +10,8 @@ from orca_python import (
 
 proc = Processor("ml")
 
-# E.g. you have you have a fleet of busses, where every bus has a particular ID and runs a particular route
+# E.g. you have you have a fleet of busses, where every bus has a particular
+# ID and runs a particular route
 route_id = MetadataField(name="route_id", description="The unique ID of the route")
 bus_id = MetadataField(name="bus_id", description="The unique ID of the bus")
 
@@ -24,6 +25,7 @@ Every30Second = WindowType(
 
 @proc.algorithm("MyAlgo", "1.0.0", Every30Second)
 def my_algorithm(params: ExecutionParams) -> StructResult:
+    """A simple algorithms that does nothing interesting"""
     route_id = params.window.metadata.get("route_id", None)
     bus_id = params.window.metadata.get("bus_id", None)
     print(route_id, bus_id)