+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+- A stubbed out `registry` module within the package that the Orca CLI generates type stubs for. This module will contain the algorithms, window types and metadata present within Orca core. The LSP that the user uses must be pointed towards the stubs generated from the CLI for this to be effective.
+
 ## [v0.10.0] - 27-09-2025
 - Bumped Orca version